# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API",
        "description": "Documentación de la API proyecto CDMI",
        "contact": {
            "email": "tucorreo@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Usuarios Registrados",
                "operationId": "600ec787a82d08ec298424fd97dd33d9",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Registrar un nuevo usuario",
                "operationId": "12af6b3d715bf3099de301b50a13d0da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellido",
                                    "fecha_nacimiento",
                                    "correo",
                                    "telefono",
                                    "contraseña",
                                    "rol"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "apellido": {
                                        "type": "string"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    },
                                    "contraseña": {
                                        "type": "string"
                                    },
                                    "rol": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/usuarios/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Iniciar sesion",
                "operationId": "27c1ad7b2797a49db722e115089c7172",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar completamente un usuario",
                "operationId": "01c6b403efd4f7869e6703e481d4a8d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellido",
                                    "fecha_nacimiento",
                                    "correo",
                                    "telefono",
                                    "contraseña",
                                    "rol"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "apellido": {
                                        "type": "string"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    },
                                    "contraseña": {
                                        "type": "string"
                                    },
                                    "rol": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado exitosamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar un usuario",
                "operationId": "811296f77edc543b83012b9455dea555",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado exitosamente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar parcialmente un usuario",
                "operationId": "39bcc7bf35ca466d3a293e87a425c665",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario a actualizar parcialmente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "apellido": {
                                        "type": "string"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    },
                                    "contraseña": {
                                        "type": "string"
                                    },
                                    "rol": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado parcialmente"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                }
            }
        }
    },
...
}